requirements.txt
tomli>=2.0.0
beautifulsoup4>=4.9.3
nltk>=3.6.2
faiss-cpu>=1.7.4
numpy>=1.23.0
tqdm>=4.65.0
requests>=2.28.0
bert-score>=0.3.13
gensim>=4.0.0
requests>=2.28.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
gunicorn==20.1.0
torch>=2.0.0
transformers>=4.30.0 